refactor(dpp): rename value_balance to unshielding_amount (u64) in Unshield and ShieldedWithdrawal
Rename `value_balance: i64` to `unshielding_amount: u64` since these
transitions always have positive outflow. Simplifies validation (== 0
instead of <= 0) and makes the type match semantics. Also adds TODO
comments on `calculate_min_required_fee` returning 0 for shielded types,
and replaces todo!() panics in path_elements display with simple strings.

+++ packages/rs-dpp/src/shielded/builder/shielded_withdrawal.rs
@@ -99,7 +99,7 @@ pub fn build_shielded_withdrawal_transition<P: OrchardProver>(
 
     ShieldedWithdrawalTransition::try_from_bundle(
         sb.actions,
-        sb.value_balance,
+        sb.value_balance as u64,
         sb.anchor,
         sb.proof,
         sb.binding_signature,
